Revert "Separate config and security file event"
This reverts commit 48da5792c11c0c77f4fe7d6ad33f1f7d2ca04e42.
