+++ kafka-connect-s3/src/main/java/io/confluent/connect/s3/file/KafkaFileEventConfig.java
@@ -26,7 +26,10 @@ public class KafkaFileEventConfig extends AbstractFileEventConfig {
 
   private String topicName;
   private String bootstrapServers;
+  private String securityProtocol;
   private String schemaRegistryUrl;
+  private String saslMecanism;
+  private String saslJaasConfig;
 
   /** empty constructor for jackson */
   public KafkaFileEventConfig() {
@@ -35,10 +38,16 @@ public class KafkaFileEventConfig extends AbstractFileEventConfig {
   public KafkaFileEventConfig(
       String topicName,
       String bootstrapServers,
-      String schemaRegistryUrl) {
+      String schemaRegistryUrl,
+      String securityProtocol,
+      String saslMecanism,
+      String saslJaasConfig) {
     this.topicName = topicName;
     this.bootstrapServers = bootstrapServers;
     this.schemaRegistryUrl = schemaRegistryUrl;
+    this.securityProtocol = securityProtocol;
+    this.saslMecanism = saslMecanism;
+    this.saslJaasConfig = saslJaasConfig;
   }
 
   @Override
@@ -55,6 +64,15 @@ public class KafkaFileEventConfig extends AbstractFileEventConfig {
     sb.append("\"topic_name\": \"").append(topicName).append('"');
     sb.append(", \"bootstrap_servers\": \"").append(bootstrapServers).append('"');
     sb.append(", \"schema_registry_url\": \"").append(schemaRegistryUrl).append('"');
+    if (securityProtocol != null) {
+      sb.append(", \"security_protocol\": \"").append(securityProtocol).append('"');
+    }
+    if (saslMecanism != null) {
+      sb.append(", \"sasl_mecanism\": \"").append(saslMecanism).append('"');
+    }
+    if (saslJaasConfig != null) {
+      sb.append(", \"sasl_jaas_config\": \"").append(saslJaasConfig).append('"');
+    }
     sb.append('}');
     return sb.toString();
   }
@@ -64,10 +82,21 @@ public class KafkaFileEventConfig extends AbstractFileEventConfig {
     Properties prop = new Properties();
     prop.setProperty("key.serializer", KEY_SERIALIZER);
     prop.setProperty("value.serializer", VALUE_SERIALIZER);
+    prop.setProperty("auto.create.topics.enable", "true");
     // mandatory
     prop.setProperty("bootstrap.servers", bootstrapServers);
     prop.setProperty("topic.name", topicName);
     prop.setProperty("schema.registry.url", schemaRegistryUrl);
+    // optional
+    if (saslMecanism != null) {
+      prop.setProperty("sasl.mechanism", saslMecanism);
+    }
+    if (securityProtocol != null) {
+      prop.setProperty("security.protocol", securityProtocol);
+    }
+    if (saslJaasConfig != null) {
+      prop.setProperty("sasl.jaas.config", saslJaasConfig);
+    }
     return prop;
   }
 
@@ -83,4 +112,15 @@ public class KafkaFileEventConfig extends AbstractFileEventConfig {
     return bootstrapServers;
   }
 
+  public String getSecurityProtocol() {
+    return securityProtocol;
+  }
+
+  public String getSaslMecanism() {
+    return saslMecanism;
+  }
+
+  public String getSaslJaasConfig() {
+    return saslJaasConfig;
+  }
 }
