+++ kafka-connect-s3/src/main/java/io/confluent/connect/s3/file/KafkaFileEventProvider.java
@@ -21,17 +21,13 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.joda.time.DateTime;
 
-import java.util.Properties;
-
 public class KafkaFileEventProvider extends FileEventProvider {
   private final KafkaFileEventConfig kafkaConfig;
-  private final KafkaFileEventSecurity kafkaSecurity;
 
-  public KafkaFileEventProvider(String configJson, String securityJson, boolean skipError) {
-    super(configJson, securityJson, skipError);
+  public KafkaFileEventProvider(String configJson, boolean skipError) {
+    super(configJson, skipError);
     this.kafkaConfig =
-        AbstractFileEventConfig.fromJsonString(configJson, KafkaFileEventConfig.class);
-    this.kafkaSecurity = AbstractFileEventConfig.fromJsonString(securityJson, KafkaFileEventSecurity.class);
+        KafkaFileEventConfig.fromJsonString(configJson, KafkaFileEventConfig.class);
   }
 
   @Override
@@ -55,11 +51,8 @@ public class KafkaFileEventProvider extends FileEventProvider {
             formatDateRFC3339(currentTimestamp),
             recordCount,
             formatDateRFC3339(eventDatetime));
-    Properties combinedProperties = new Properties();
-    combinedProperties.putAll(kafkaConfig.toProps());
-    combinedProperties.putAll(kafkaSecurity.toProps());
     try (final Producer<String, SpecificRecord> producer =
-        new KafkaProducer<>(combinedProperties)) {
+        new KafkaProducer<>(kafkaConfig.toProps())) {
       producer.send(
           new ProducerRecord<>(kafkaConfig.getTopicName(), key, value),
           (event, ex) -> {
